Consolidate modeling structs to a single module
